requirements.txt
asgiref==3.2.7
Django==3.0.5
django-widget-tweaks==1.4.8
mypy==0.971
mypy-extensions==0.4.3
mysqlclient==2.1.1
numpy==1.23.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.2.0
pygame==2.1.2
pytz==2020.1
sqlparse==0.3.1
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
jsonfield==3.1.0